本地更新cpp文件
修改定义内容

+++ app/src/main/java/com/luck/pictureselector/newtestfile.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-    float d,e,f;
+    int d,e,f;
     int a,b,c;
     printf("hello world!\n");
     return 0;
